requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
gunicorn==23.0.0
python-dotenv==1.0.1
anthropic>=0.7.0
firebase-admin==5.0.0
google-cloud-firestore==2.3.4
google-cloud-storage==2.14.0
requests==2.26.0
setuptools>=65.5.1
flask-cors==4.0.0
google-cloud-logging==3.8.0
pyOpenSSL==23.2.0
cryptography==41.0.4